+++ src/Row.php
@@ -17,19 +17,14 @@ namespace MarcinOrlowski\TextTable;
 use Lombok\Getter;
 use Lombok\Setter;
 use MarcinOrlowski\TextTable\Exceptions\DuplicateColumnKeyException;
-use MarcinOrlowski\TextTable\Traits\ArrayAccessTrait;
-use MarcinOrlowski\TextTable\Traits\IteratorAggregateTrait;
 
 /**
  * @method CellsContainer getContainer()
  */
 #[Getter]
 #[Setter]
-class Row extends \Lombok\Helper implements ContainerContract
+class Row extends BaseContainer
 {
-    use ArrayAccessTrait;
-    use IteratorAggregateTrait;
-
     /**
      * @param array|null $cells Optional list of cells to be added to the newly created row.
      *